Overhauled Camera::flushOutput() so that now we have per exposure a list of stars and their positions.

+++ include/Camera.h
@@ -78,7 +78,10 @@ class Camera : public HDF5Writer
         PointSpreadFunction *psf;
         Polynomial1D polynomial;
 
-        map<unsigned int, map<double, array<double, 6>>> detectedStarInfo;  // detectedStarInfo[starID][startTime] contains the values (xFPmean, yFPmean, sumFlux, Ndetections} values
+        // detectedStarInfo[startTime][starID] contains the values 
+        //    (xFPmean, yFPmean, rowPixMean, colPixmean, sumFlux, Ndetections} values
+
+        map<double, map<unsigned int, array<double, 6>>> detectedStarInfo;
 
 };
